requirements.txt
numpy
pillow
matplotlib
torch
torchvision
opencv-python
tdqm
